✨ feat(watchMode): add unified diff strategy implementation
- create unifiedDiffStrategy.ts to handle parsing and applying unified diffs
- implement various utility functions and classes for diff handling
- include error handling and logging for debugging purposes
